@@ -20,11 +20,11 @@
     ],
     "homepage": "https://docs.sonata-project.org/projects/SonataBlockBundle",
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": "^8.0",
         "doctrine/collections": "^1.6",
         "doctrine/common": "^3.0",
         "psr/container": "^1.0 || ^2.0",
-        "psr/log": "^1.1 || ^2.0 || ^3.0",
+        "psr/log": "^2.0 || ^3.0",
         "sonata-project/form-extensions": "^1.4",
         "symfony/asset": "^5.4 || ^6.0",
         "symfony/config": "^5.4 || ^6.0",
