@@ -106,7 +106,7 @@ final class ContainerBlockService extends AbstractBlockService implements Editab
     private function getDecorator(string $layout): array
     {
         $key = '{{ CONTENT }}';
-        if (false === strpos($layout, $key)) {
+        if (!str_contains($layout, $key)) {
             return [];
         }
 
